requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
pdf2image
Pillow
numpy
pandas